backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pdfplumber==0.11.4
pandas==3.0.2
aiosqlite==0.20.0
python-multipart==0.0.18
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
